[cacerts] Added cwd to copy from omnibus cache rather than where fetcher tries to unpack it
There is a bug in omnibus-ruby that may or may not have been fixed.
Since the source url does not point to an archive, omnibus-ruby tries to copy
cacert.pem into the project working directory. However, it fails and copies to
'/var/cache/omnibus/src/cacerts-2012.12.19\' instead There is supposed to be
a fix in omnibus-ruby, but under further testing, it was unsure if the fix
worked. Rather than trying to fix this now, we're filing a bug and copying the
cacert.pem directly from the cache instead.