Remove the mysql2 gem software definition
This appears to have been explicitly added for MySQL support in
Omnibus builds of Chef Server (Open Source and Enterprise).  Since
MySQL is no longer supported, removal seems prudent.

This was spurred by the fact that the recent 9.0.0 release of the
rake-compiler gem requires rubygems 1.8.25 or greater; use of this
definition as is would require choices that would potentially affect
other components of a Ruby-heavy Omnibus build.