[perl] generate an Omnibus-friendly CPAN config
This will ensure CPAN can use the Omnibus cache directory for
building.  Invocations of Perl will need to be updated to pass the
`-j` option for this to have any effect, though.

This PR configures the CPAN home to
/var/cache/omnibus/cache/cpan. Previously the CPAN home leaked out
into $HOME/.cpan and could cause build slaves to run out of inodes as
this directory was never cleaned up!

(@schisamo did the work; I'm just the messenger.)