Switch ftp:// URLs to http://
FTP requires a range of possible ports to be open for transferring
files, which makes it difficult to use in network environments with
strict firewall rules. HTTP, on the other hand, uses a single port and
doesn't require any special firewall rules.

There are four packages that use ftp:// for their URL. Since all of
them have an equivalent http:// URL, use that instead.