Add GMP, GNU Multiple Precision Arithmetic library
Create a software configuration for GMP v4.1.3.  I used the GDBM configuration
as a starting point which is how the "platform == freebsd" hunk found it's way
in here, but I don't actually have a FreeBSD machine to test that on.

Also, Ruby 2.1 depends on GMP (see: https://bugs.ruby-lang.org/issues/8796)
so mark the dependency when building Ruby v2.1.1.

Signed-off-by: Kent R. Spillner <kspillner@acm.org>