Lock down versions of all ChefDK's auxiliary gems
Currently when we create a build of ChefDK we get the latest released 
version of every auxiliary gem. This is dangerous at best. This change
locks down the versions.

I also wonder if we should just put these in the ChefDK gemspec OR 
Gemfile. If we decide to create `chef` subcommands that rely on any of
these tools as a library we'll have to (I'm thinking `chef lint`).