add comment that this software defn is deprecated
i'll add a bit more background in this commit message, with the stated
qualifier for anyone in the public (non-Chef Software, Inc. employees)
stumbling upon this that THIS IS NOT LEGAL ADVICE, IANAL, etc.

discussed with Adam J about dynamically linking vs statically linking
this due to LGPL concerns and we believe that there's no concerns over
statically linking libgcc, and that we've been cargo culting this
particular solution for a few years for no reason.