Revert 2a89c31
The change to the jre config broke builds and upon further review, the
way in which jre versions are hard-coded will prevent projects wishing
to override the default_version provided here (which doesn't align with
the hard-coded version causing some of the breakage issues).

So for now, we're reverting this to get back to a working state.