Remove .git dir from embedded/apps
The presence of a .git directory causes git to treat these directories
as submodules, which causes the content to *not* get cached by the
omnibus git cache, which in turn causes cached builds to have empty
directories where the code was supposed to be. There is some benefit to
keeping the .git directories around, but not enough to justify a more
involved fix/workaround. So, blow the .git directories away during the
build.