Use `cpanminus` to install Perl module deps
We’ll also ensure we don’t run tests when installing CPAN modules. Go
read this if you need a reason why this is insane:
http://www.openswartz.com/2012/01/31/stop-running-tests-on-install/

We’ve hit issues where a modules test’s fail (and thus the module is NOT
installed) and the default `cpan` client continues along and exits with a
zero status!