compile ncurses with xlc, remove libtool dep
This omits building the ncurses c++ bindings on AIX, which we do not
link against in the client.  The Makefiles for the c++ bindings
could probably be fixed since the root cause is that it uses ar instead
of xlc with -qmkshrobj like the rest of them and probably just needs
fixing, but tl;dr: we don't need it.

Also disables the Ada bindings as a bonus because I don't think we
have any Ada code.