Take advantage of globbing in windows `gem build` command
This form of the `gem build` command will build a gemspec named
`chef-windows.gemspec` or `chef-x86-mingw32.gemspec`. Previously We
used `File.exist?` guards to determine which gemspec to build. The
problem with this approach is the guards are evaluated at software
definition load time instead of the desired build execution time. If
the Chef source code does not exist at the beginning of the build, the
Chef Windows gem will not be built as both guards will evaluate to
false. This can occur when the build cache was intentionally expired
(EXPIRE_CACHE parameter) OR a build is being executed on a freshly
provisioned Windows build slave. The `File.exist?` guards were also
incorrectly evaluated against the versio of the Chef source code from
the previous build!