Patch libffi to only use the {prefix}/lib dir
This patch is cross platform and ensures the libraries are built
and placed in the correct location initially, instead of
attempting to move them after the fact. Now all libffi targets
are installed in {prefix}/lib and {prefix}/lib64 is not used
when a command line flag to configure is specified. This may
be overridden for cross-platform builds if necessary. Tested
on muiltiple 64 and 32 platforms (including ones specifically
worked around previously)