Remove extended attribute support from patch
When building omnibus build which include patch, extended
attribute support is enabled by default when support libraries
are detected. There is no cross-platform support for these in
a single library and adding whitelist exceptions for every
possible platform seems like an anti-pattern. Also, for the type
of work we do with patch (just source code) extended attribute
support is not necessary so I've disabled it.