Refactor ruby and rubygems to allow for customization
This is a major change in how ruby and rubygems can be brought into any
omnibus software or project definition. Previously, you could use ruby
and ruby-windows to request a version of ruby on either of these
platforms and you would manually do something like

  if windows?
    override :'ruby-windows', version: something
    dependency 'ruby-windows'
  else
    override :ruby, version: something
    dependency 'ruby'
  end

With the proliferation of various combinations of how ruby is built and
where the compiler toolchain comes from, this is now changing. From now
on, any software that needs ruby may simply say

  dependency "ruby"

To override which version of ruby you wish to install, your project must
first specify how it would like to get its ruby. The version string of
the ruby software definition now acts as that label. You then have to
specify the actual version of ruby that you'd like by overriding the
version of a child ruby definition (usually ruby-<thelabel>).

e.g.

  override :ruby, version: "compiled"
  override :"ruby-compiled", version: "2.1.7"

This requests ruby to be compiled on all platforms using whatever system
compiler that platform provides on the path. See the build-essentials
cookbook for more info on what's provided on windows.

  if windows?
    override :ruby, version: "rubyinstaller"
    override :"ruby-rubyinstaller", version: "2.1.6"
  end

This requests the ruby installer binary package to be fetched and
unzipped on windows. Note that no devkit is provided. Any gems built
will use the system compiler available on the path.

As part of this change, no devkit is installed on windows by default. If
you want to vendor a compiler, you'll need to ask for that. The
rubygems-native software definition is a metaproject that configures
that. Once again, the version string for rubygems-native is simply the
flavor of toolchain you wish to configure for your ruby. Changing the
version of a vendored tool is done by overriding the version string of
the software definition that handles that particular flavor of
toolchain.

e.g.

  override :'rubygems-native', version: "system-default"

This requests that all gem compilation use the system compiler available
on the path on all platforms.

  override :'rubygems-native', version: "mingw" if windows?

This will create a dependency on the mingw toolchain which brings down
an appropriate version of gcc, msys and other utilities. It also
registers said compiler with ruby (a la dk.rb from devkit).

One thing to remember is that now, any software definition that installs
a gem (using gem or bundler...  it doesn't matter) and has a native gem
in the transitive closure of its dependencies must declare
"rubygems-native" as a dependency (instead of simply "rubygems"). This
will ensure that the correct compiler is installed and available in the
path before a gem is compiled. If instead, you know that a native gem in
your transitive closure of dependencies is better handled by another
omnibus software definition, you may instead just include that as a
dependency. For example, if you need nokogiri but don't otherwise need
any other native gems, you can specify:

dependency "rubygems"
dependency "nokogiri"