Keep the ruby-windows-devkit dependencies out of ruby
The reason is that ruby does not depend of devkit. Our
devkit implementation depends on ruby as it needs it
for the install script phase. Removing this allows
us to add ruby as a dependency of ruby-windows-devkit
and have it automagically get the right ruby that was
compiled in our software definition or the one that
came from rubyinstaller