Make libzmq not assume omnibus cache directory structure
The libzmq software definitions assumes that the omnibus source/cache
directory structure (like the fact that it used to unzip things
directly into the cache dir).  Recently, the layout changed to introduce
per-package directories into which things were unzipped.  The correct
way to write the definition is to use relative_path to define the
directory you want to work under and simply assume that all the DSL
methods get run in that directory.