[pip deps] Remove `-I` option (#77)
The option makes pip ignore the pip pkgs that are already installed.

Removing it ensures that pip never installs the same package twice with
different versions (in which case we can't reliably predict which
version will actually be loaded by python).

This would happen for `requests`.

NB: we still don't fully guarantee that the requested version of a pip
package will be the one that's in the final package. For instance in
this (made-up) scenario:
1. install requests 2.5.0
2. install docker-py 1.8.1 (which requires requests >2.5.2)

The install of `docker-py` would remove the existing requests and
replace it with a higher version of requests, so the final package would
end up including requests 2.5.3 for instance.

We still need some after-the-fact checks to ensure that we are indeed
packaging the right versions of the pip deps we've pinned. These
checks are currently done in our internal dd-agent-testing job.