Import the veil-gem software def from chef-server
We likely want to use this in a number of add-ons as well so it would
be good to share it.  In the long run, it would be better to not have
this as an omnibus software definition at all. Instead, we should pull
it in via a Gemfile in the appropriate apps.  But, we aren't there yet.

Signed-off-by: Steven Danna <steve@chef.io>