remove iconv from git
chefdk pulls in git which pulls in iconv, so need to remove this as well

it looks like this breaks charset conversion in git-mailinfo
functionality.  we're not trying to replace or provide a distro git
level of support here, and nothing in the chef-dk workflow usage of
git should be broken by this.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>