Revert "remove iconv from git"
This reverts commit 1e8890e700445cbbc0408270133375ff45bc8551.

chef-dk doesn't actually use this definition, it forked it and uses
its own.  this also doesn't actually work and compiling git on
mac needs libiconv and there's no way to turn it off.  we might
think about not shipping libiconv and only allow the mac build to
link against libiconv in the operating system.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>