Relocating Windows OpenSSL DLLs
When OpenSSL starts in FIPS mode it does a self-check. Part of this
self-check is looking at the openssl code loaded into memory and
ensuring it matches a known checksum. If Windows relocates the OpenSSL
DLLs because of an already-loaded image-base conflict then this
self-check will fail. So we hard code the image base for these items on
Windows to prevent them getting relocated

Signed-off-by: tyler-ball <tyleraball@gmail.com>
Signed-off-by: rmoshier <radler@chef.io>