postgresql: add versioned and bin-only versioned variants
This is for upgrading PG: the project depends on both

    postgres-versioned@9.6.3

and

    postgres-bin-versioned@9.2.21

thus you end up with the "full" postgres in

    /opt/PROJECT/embedded/postgresql/9.6/{share,bin,lib}

and a more minimal, but functional, one in

    /opt/PROJECT/embedded/postgresql/9.2/{bin,lib}

This allows for smooth upgrades, since pg_upgrade (the in-place upgrade
utility shipped with PG) requires the binaries of the old version to be
present.

Also, it circumvents omnibus' restriction of not being able to depend on
one software definition _twice_

I couldn't find a better way to _not_ put the same code in three
different software definitions, but I'd be glad to learn of one! :)

Signed-off-by: Stephan Renatus <srenatus@chef.io>