[libiconv] Move to 1.15
This removes the patches we have in place.

The git removal patch appears to be no longer needed based on the
following commits in the upstream gnulib:

  66712c23388e93e5c518ebc8515140fa0c807348
  d6176702c3ffb6c56e6620397470e1b3bd540b10

The ppc64le patch may still be needed; however, it doesn't apply and
the new configure scripts *appear* to handle ppc64le now.

Signed-off-by: Steven Danna <steve@chef.io>