Avoid Lists::MoreUtils::XS because of EL 5 failures
Lists::MoreUtils is a transitive dependency of sqitch.  In the default
configuration, it depends on Lists::MoreUtils::XS.

The Lists::MoreUtils::XS perl module fails to install correctly. The
issue is bad assumptions in its configuration script leading to a
duplicate definition of ssize_t leading to a failed compilation.

The PUREPERL_ONLY option avoids this module by substituting a pure
perl implementation.

Signed-off-by: Steven Danna <steve@chef.io>