Update PostgreSQL versions
There are new releases of all supported versions of PostgreSQL for
CVE-2018-10915 and CVE-2018-10925.

I've gone through and added the latest for each of 9.* releases, reordered the definitions by version and noted deprecation dates for each version.

Future work should include updating the default (it's long deprecated, but changing that will break people), and ripping out EoL versions.

I didn't add anything for postgres10, we should add a separate postgres10 definition for that.

Signed-off-by: Mark Anderson <mark@chef.io>