Add Ruby 2.5.3/2.4.5. Make 2.5.3 the default
Also remove the super old and very much EOL 2.2 releases. We have some
EOL apps we *may* want to rebuild that depend on some of these 2.2
releases so don't entirely remove it. Nothing at Chef is unpinned so
bumping the default to 2.5.3 is a no-op, but should be done for any
project getting started since 2.2 is EOL.

Signed-off-by: Tim Smith <tsmith@chef.io>