get rubygems' submodules else source builds fail
rubygems/rubygems started vendoring bundler as a submodule[1]
in 2016. Possibly we don't have any projects using RubyGems from
a git reference since then. But in trying to do so, builds of rubygems
from a get ref error out with:

    ERROR:  While executing gem ... (Errno::ENOENT)
    No such file or directory @ dir_chdir - bundler/lib

So, as much as submodules sadden me, to run rubygems' setup.rb,
we need to initialize and update the repo's submodules.

[1] https://github.com/rubygems/rubygems/commit/aa83d3799579b8714b1f9c8c78afdb4e176bba58#diff-db852b71825b722d7e0dcb631d77414d

Signed-off-by: Robb Kidd <robb@thekidds.org>