Revert "Merge pull request #1071 from mattray/make_4_2_1"
This reverts commit 5fa35959f6efb0a5a745c46e4b0431a2826934da, reversing
changes made to ed46161c881e2a76200f08e2b7b1b8600e958852.

The sed command in question does not work on all supported systems, and
the conditions for when this command should be run are unclear. This
fix was introduced for an issue in glib-2.27, but the effect of this
patch on other versions of glibc is unknown. Also, we want to use the
`patch` DSL when modifying source files.

Signed-off-by: Tom Duffield <tom@chef.io>