Remove support for Ruby 2.4 / 2.5 fully
We recently updated all of these defintions to utilize the bundler that
is built into Ruby itself. That means we dropped support for Ruby < 2.6
where bundler isn't built-in. The current state of this repo makes it
appear that we still support Ruby 2.4/2.5 when we actually don't. Users
(including Chef) who still build legacy apps on 2.4/2.5 need to use an
older commit of this repo.

Signed-off-by: Tim Smith <tsmith@chef.io>