Exclude the debug group to avoid double pry from Ohai
Ohai comes in as a git repo in chef/chef which plows right past the
gemfile.lock in the appbundle and it has a debug group that includes
pry. This is resulting in a newer version of pry coming in via ohai than
the rest of chef and that's causing double pry problems.

Signed-off-by: Tim Smith <tsmith@chef.io>