WIP fixes for openssl on M1
This provides a patch which adds awareness of the MacOS M1 platform to
1.0.2x. This patch is adapted from this repo [1] which is based on
1.0.1j. I did a manual diff between this repo and the 1.0.1j to see what
changes were really needed.

In addition, it seems `-Qunused-arguments` needs to be enabled to work
aroudn clang compiler issues.

Currently I'm running into following compile issue that I haven't found
a solution for:

```
1 warning generated.
In file included from <built-in>:382:
<command line>:4:19: warning: ISO C99 requires whitespace after the macro name [-Wc99-extensions]
```

[1] https://github.com/carsenk/openssl10-darwin-arm

Signed-off-by: Lance Albertson <lance@osuosl.org>