Add ncurses dependency to ruby's software definition
we build omnibus packages on freebsd 11 and use the packages on freebsd 11, 12 and 13.
the ruby executable has been linking to freebsds system's ncurses library files.
freebsd 13 system's ncurses library files have a different name than freebsd 11 and 12
which causes the ruby executable to fail.
adding ncurses as a dependency for freebsd prevents the ruby executable from linking to the
system's ncurses library files thereby allowing the package built on freebsd 11 to work on freebsd 13.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>