Revert "Use smaller docker image"
This reverts commit ea9ff306ec154e6372454ae69048e75ad361397b.

We deployed new AMIs to the buildkite queues that run the verify
pipeline and unfortunately they have an issue that is preventing us from
using the ubuntu:18.04 docker image.
We have a fix ready but we need some time
to resolve some deployment issues so in the meantime we are going to use
chefes/releng-base again.

Signed-off-by: Jeremiah Snapp <jeremiah@chef.io>