Use disable-dynamicbase and fixed DLL locations
disable-dynamicbase seems to be the key for preventing "Fingerprint does
not match" which is likely due to ASLR (Address Space Layout
Randomization) on Windows, which I suspect is related to the switch to
UCRT. Of note is that I also had to add the crypt lib / 0x63000000 line
in order for OpenSSL fips_mode to enable successfully.

Not 100% certain that mingw64 and mingw sections need to coexist, but
the platform now passed to ./Configure for openssl-fips 2.0.16 source is
now "mingw64". The "mingw" sections should minimize impact to pre-UCRT
versions of ruby builds.

Signed-off-by: Thomas Powell <powell@progress.com>